refactor(api): modularizar enpoints y modelos de datos de information sources y stats
