+++ newsradar_api/app/api/router.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
 from datetime import datetime, timezone
-from typing import Dict, List, Optional
+from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException, Response
 from fastapi.security import HTTPBearer
-from pydantic import BaseModel, Field, HttpUrl
 
 from app.schemas.roles import Role
 from app.stores.memory import users_store, roles_store
@@ -21,6 +20,7 @@ from app.stores.memory import (
     categories_store,
     notifications_store,
     rss_channels_store,
+    information_sources_store,
 )
 
 from app.utils.user_utils import ensure_role_ids_exist, sanitize_user
@@ -31,6 +31,8 @@ from .routes.roles import roles_router
 from .routes.alerts import api_alerts_router
 from .routes.notifications import notifications_router
 from .routes.categories import categories_router
+from .routes.information_sources import information_sources_router
+from .routes.stats import stats_router
 
 api_router = APIRouter()
 api_router.include_router(api_auth_router)
@@ -39,52 +41,12 @@ api_router.include_router(roles_router)
 api_router.include_router(api_alerts_router)
 api_router.include_router(notifications_router)
 api_router.include_router(categories_router)
+api_router.include_router(information_sources_router)
+api_router.include_router(stats_router)
 
 API_PREFIX = "/api/v1"
 security = HTTPBearer(auto_error=False)
 
-
-class Metric(BaseModel):
-    name: str = Field(..., min_length=1, max_length=100)
-    value: float
-
-
-class InformationSourceBase(BaseModel):
-    name: str = Field(..., min_length=1, max_length=120)
-    url: HttpUrl
-
-
-class InformationSourceCreate(InformationSourceBase):
-    pass
-
-
-class InformationSourceUpdate(BaseModel):
-    name: Optional[str] = Field(None, min_length=1, max_length=120)
-    url: Optional[HttpUrl] = None
-
-
-class InformationSource(InformationSourceBase):
-    id: int
-
-
-class StatsBase(BaseModel):
-    metrics: List[Metric] = Field(default_factory=list)
-
-
-class StatsCreate(StatsBase):
-    pass
-
-
-class StatsUpdate(BaseModel):
-    metrics: Optional[List[Metric]] = None
-
-
-class Stats(StatsBase):
-    id: int
-
-
-information_sources_store: Dict[int, InformationSource] = {}
-stats_store: Dict[int, Stats] = {}
 
 counters = {
     "roles": 1,
@@ -181,95 +143,6 @@ def health() -> dict:
 
 
 @api_router.get(
-    f"{API_PREFIX}/information-sources",
-    response_model=List[InformationSource],
-    tags=["information-sources"],
-)
-def list_information_sources(
-    _: UserInDB = Depends(get_current_user),
-) -> List[InformationSource]:
-    return list(information_sources_store.values())
-
-
-@api_router.post(
-    f"{API_PREFIX}/information-sources",
-    response_model=InformationSource,
-    status_code=201,
-    tags=["information-sources"],
-)
-def create_information_source(
-    payload: InformationSourceCreate,
-    _: UserInDB = Depends(get_current_user),
-) -> InformationSource:
-    source_id = next_id("information_sources")
-    source = InformationSource(id=source_id, **payload.model_dump())
-    information_sources_store[source_id] = source
-    return source
-
-
-@api_router.get(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    response_model=InformationSource,
-    tags=["information-sources"],
-)
-def get_information_source(
-    source_id: int, _: UserInDB = Depends(get_current_user)
-) -> InformationSource:
-    source = information_sources_store.get(source_id)
-    if not source:
-        raise HTTPException(
-            status_code=404, detail="Fuente de información no encontrada"
-        )
-    return source
-
-
-@api_router.put(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    response_model=InformationSource,
-    tags=["information-sources"],
-)
-def update_information_source(
-    source_id: int,
-    payload: InformationSourceUpdate,
-    _: UserInDB = Depends(get_current_user),
-) -> InformationSource:
-    source = information_sources_store.get(source_id)
-    if not source:
-        raise HTTPException(
-            status_code=404, detail="Fuente de información no encontrada"
-        )
-    updated = source.model_copy(update=payload.model_dump(exclude_unset=True))
-    information_sources_store[source_id] = updated
-    return updated
-
-
-@api_router.delete(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["information-sources"],
-)
-def delete_information_source(
-    source_id: int, _: UserInDB = Depends(get_current_user)
-) -> None:
-    if source_id not in information_sources_store:
-        raise HTTPException(
-            status_code=404, detail="Fuente de información no encontrada"
-        )
-
-    channel_ids = [
-        channel.id
-        for channel in rss_channels_store.values()
-        if channel.information_source_id == source_id
-    ]
-    for channel_id in channel_ids:
-        rss_channels_store.pop(channel_id, None)
-
-    information_sources_store.pop(source_id, None)
-
-
-@api_router.get(
     f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels",
     response_model=List[RSSChannel],
     tags=["rss-channels"],
@@ -361,58 +234,3 @@ def delete_source_channel(
     ensure_information_source_exists(source_id)
     ensure_rss_for_source(source_id, channel_id)
     rss_channels_store.pop(channel_id, None)
-
-
-@api_router.get(f"{API_PREFIX}/stats", response_model=List[Stats], tags=["stats"])
-def list_stats(_: UserInDB = Depends(get_current_user)) -> List[Stats]:
-    return list(stats_store.values())
-
-
-@api_router.post(
-    f"{API_PREFIX}/stats", response_model=Stats, status_code=201, tags=["stats"]
-)
-def create_stats(
-    payload: StatsCreate, _: UserInDB = Depends(get_current_user)
-) -> Stats:
-    stats_id = next_id("stats")
-    stats = Stats(id=stats_id, **payload.model_dump())
-    stats_store[stats_id] = stats
-    return stats
-
-
-@api_router.get(
-    f"{API_PREFIX}/stats/{{stats_id}}", response_model=Stats, tags=["stats"]
-)
-def get_stats(stats_id: int, _: UserInDB = Depends(get_current_user)) -> Stats:
-    stats = stats_store.get(stats_id)
-    if not stats:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-    return stats
-
-
-@api_router.put(
-    f"{API_PREFIX}/stats/{{stats_id}}", response_model=Stats, tags=["stats"]
-)
-def update_stats(
-    stats_id: int, payload: StatsUpdate, _: UserInDB = Depends(get_current_user)
-) -> Stats:
-    stats = stats_store.get(stats_id)
-    if not stats:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-
-    updated = stats.model_copy(update=payload.model_dump(exclude_unset=True))
-    stats_store[stats_id] = updated
-    return updated
-
-
-@api_router.delete(
-    f"{API_PREFIX}/stats/{{stats_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["stats"],
-)
-def delete_stats(stats_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if stats_id not in stats_store:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-    stats_store.pop(stats_id, None)
